+++ CHANGELOG.md
@@ -459,5 +459,6 @@ v912
 * added NMEA protocol - now this firmware can be used without any FC
 
 v913
-*added new panel - motors output
+* added new panel - motors output
+* fixed last CT bugs